requirements.txt
numpy==2.1.3
openai==1.55.3
pandas==2.2.3
parmap==1.7.0
python-dotenv==1.0.1
scikit_learn==1.5.2
